# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing universities, departments, teams, proposals, and reviews.",
        "title": "University Project Hub API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/proposals": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve proposals with optional filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proposals"
                ],
                "summary": "Get proposals",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Proposal status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Department ID",
                        "name": "department_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.Proposal"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new proposal ID with version 1. Team is optional at this stage.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proposals"
                ],
                "summary": "Create a new proposal draft",
                "parameters": [
                    {
                        "description": "Proposal details",
                        "name": "proposal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/proposals.SaveProposalRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Proposal"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/proposals/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a specific proposal by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proposals"
                ],
                "summary": "Get proposal by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Proposal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Proposal"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "If Draft: updates existing. If Rejected/Revision: creates new version.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proposals"
                ],
                "summary": "Update proposal or create revision",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Proposal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Proposal details",
                        "name": "proposal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/proposals.SaveProposalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Proposal"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a proposal if it is in Draft status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proposals"
                ],
                "summary": "Delete a proposal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Proposal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/proposals/{id}/submit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Locks proposal and sends to Admin. Requires Finalized Team.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proposals"
                ],
                "summary": "Submit proposal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Proposal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Team ID Confirmation",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/proposals.SubmitProposalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/proposals/{id}/versions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a specific proposal by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proposals"
                ],
                "summary": "Get proposal by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Proposal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.ProposalVersion"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Department": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "e.g., CSE, SE",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "university": {
                    "$ref": "#/definitions/domain.University"
                },
                "university_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "domain.Proposal": {
            "type": "object",
            "properties": {
                "advisor_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/enums.ProposalStatus"
                },
                "team": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Team"
                        }
                    ]
                },
                "team_id": {
                    "description": "⚠️ Changed to pointer to allow NULL",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "versions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ProposalVersion"
                    }
                }
            }
        },
        "domain.ProposalVersion": {
            "type": "object",
            "properties": {
                "abstract": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "expected_timeline": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_approved": {
                    "type": "boolean"
                },
                "methodology": {
                    "type": "string"
                },
                "objectives": {
                    "type": "string"
                },
                "problem_statement": {
                    "type": "string"
                },
                "proposal_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "version_number": {
                    "type": "integer"
                }
            }
        },
        "domain.Team": {
            "type": "object",
            "properties": {
                "advisor_id": {
                    "description": "Admin assigns this later",
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "department_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_finalized": {
                    "description": "The \"Lock\"",
                    "type": "boolean"
                },
                "members": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.TeamMember"
                    }
                },
                "name": {
                    "type": "string"
                },
                "proposals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Proposal"
                    }
                }
            }
        },
        "domain.TeamMember": {
            "type": "object",
            "properties": {
                "invitation_status": {
                    "$ref": "#/definitions/enums.InvitationStatus"
                },
                "role": {
                    "description": "'leader', 'member'",
                    "type": "string"
                },
                "team_id": {
                    "type": "integer"
                },
                "user": {
                    "description": "Preload User details for UI",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "domain.University": {
            "type": "object",
            "properties": {
                "academic_year": {
                    "type": "string"
                },
                "ai_checker_enabled": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "project_period": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "visibility_rule": {
                    "description": "private, public, restricted",
                    "type": "string"
                }
            }
        },
        "domain.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "department": {
                    "$ref": "#/definitions/domain.Department"
                },
                "department_id": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "email_verified": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "last_login_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "profile_photo": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/enums.Role"
                },
                "student_id": {
                    "type": "string"
                },
                "university": {
                    "$ref": "#/definitions/domain.University"
                },
                "university_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "enums.InvitationStatus": {
            "type": "string",
            "enum": [
                "pending",
                "accepted",
                "rejected"
            ],
            "x-enum-varnames": [
                "InvitationStatusPending",
                "InvitationStatusAccepted",
                "InvitationStatusRejected"
            ]
        },
        "enums.ProposalStatus": {
            "type": "string",
            "enum": [
                "draft",
                "submitted",
                "under_review",
                "revision_required",
                "approved",
                "rejected"
            ],
            "x-enum-varnames": [
                "ProposalStatusDraft",
                "ProposalStatusSubmitted",
                "ProposalStatusUnderReview",
                "ProposalStatusRevisionRequired",
                "ProposalStatusApproved",
                "ProposalStatusRejected"
            ]
        },
        "enums.Role": {
            "type": "string",
            "enum": [
                "student",
                "advisor",
                "admin",
                "public"
            ],
            "x-enum-varnames": [
                "RoleStudent",
                "RoleAdvisor",
                "RoleAdmin",
                "RolePublic"
            ]
        },
...
        "proposals.SaveProposalRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "abstract": {
                    "type": "string"
                },
                "expected_timeline": {
                    "type": "string"
                },
                "methodology": {
                    "type": "string"
                },
                "objectives": {
                    "type": "string"
                },
                "problem_statement": {
                    "type": "string"
                },
                "team_id": {
                    "description": "Optional",
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "proposals.SubmitProposalRequest": {
            "type": "object",
            "required": [
                "team_id"
            ],
            "properties": {
                "team_id": {
                    "type": "integer"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Error message"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}